+++ Yeti/Controllers/Article/CustomizeVC/CustomizeVC.m
@@ -50,6 +50,12 @@
     
 }
 
+- (UIModalPresentationStyle)modalPresentationStyle {
+    
+    return UIModalPresentationPopover;
+    
+}
+
 #pragma mark - Getters
 
 - (NSArray *)bodyFonts {